+++ lib/views/view_categories/view_categories.dart
@@ -88,10 +88,12 @@ class ViewCategoriesState extends ViewWidgetState<Category> {
 
   @override
   Widget getTitle() {
-    return Column(children: <Widget>[
-      ViewHeader(getClassNamePlural(), numValueOrDefault(list.length), getDescription()),
-      renderToggles(),
-    ]);
+    return ViewHeader(
+      title: getClassNamePlural(),
+      count: numValueOrDefault(list.length),
+      description: getDescription(),
+      child: renderToggles(),
+    );
   }
 
   @override